requirements.txt
appdirs==1.4.2
packaging==16.8
pynetworktables==2017.0.6
pyparsing==2.1.10
pyserial==3.2.1
six==1.10.0
